% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{access_coefficient}
\alias{access_coefficient}
\title{Starting from a list of fits, provide a function to access a coefficient given by the user as a string.}
\usage{
access_coefficient(fits, coefficient)
}
\arguments{
\item{fits}{A list of fit objected obtained by a call to \code{ppjsdm::gibbsm}.}

\item{coefficient}{A string representing a coefficient of interest.
Choice of \code{alpha1}, \code{alpha2}, ..., to show one of the short-range interaction coefficients;
\code{gamma} to show the medium-range interaction coefficient;
\code{beta1}, \code{beta2}, ... or actual name of the covariate for one of the regression coefficients.
\code{alpha} shows a facet plot of all short-range interaction coefficients, while \code{beta} does the same for the regression coefficients.}
}
\description{
Starting from a list of fits, provide a function to access a coefficient given by the user as a string.
}
\keyword{internal}
