% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{add_to_configuration}
\alias{add_to_configuration}
\title{Add a point to a configuration.}
\usage{
add_to_configuration(configuration, x, type)
}
\arguments{
\item{configuration}{Configuration.}

\item{x}{Point.}

\item{type}{Point type.}
}
\description{
Add a point to a configuration.
}
