% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{area.Disk_window}
\alias{area.Disk_window}
\title{Return the area of a disk window.}
\usage{
\method{area}{Disk_window}(w)
}
\arguments{
\item{w}{The window.}
}
\description{
Return the area of a disk window.
}
