% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Im_window.R
\name{area.Im_window}
\alias{area.Im_window}
\title{Return the area of an im window.}
\usage{
\method{area}{Im_window}(w)
}
\arguments{
\item{w}{Window.}
}
\description{
Return the area of an im window.
}
\examples{
if(require(spatstat.geom)) { # If no spatstat, im windows cannot be constructed
# Construct a window

window <- ppjsdm::Im_window(as.im(function(x, y) ifelse(x^2 + y^2 <= 1, 1, NA), W = owin()))

# Compute area of an im window

print(area(window))
}
}
