% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_generics.R
\name{area}
\alias{area}
\title{Area of a Window}
\usage{
area(w)
}
\arguments{
\item{w}{Window.}
}
\description{
Imported directly from \code{spatstat}, in order to provide methods for our window classes,
see for example \code{?area.Rectangle_window}. The documentation here is sparse, and we
refer to the \code{spatstat} documentation for more details.
}
