% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{area.Rectangle_window}
\alias{area.Rectangle_window}
\title{Return the area of a rectangle window.}
\usage{
\method{area}{Rectangle_window}(w)
}
\arguments{
\item{w}{Window.}
}
\description{
Return the area of a rectangle window.
}
\examples{
# Construct a window

window <- ppjsdm::Rectangle_window(x_range = c(0, 1), y_range = c(1, 3))

# Compute area of a rectangle window

print(area(window))
}
