% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{as.data.frame.Configuration}
\alias{as.data.frame.Configuration}
\title{Convert Configuration to a data.frame.}
\usage{
\method{as.data.frame}{Configuration}(x, ...)
}
\arguments{
\item{x}{Configuration.}

\item{...}{Unused.}
}
\description{
Convert Configuration to a data.frame.
}
\examples{
set.seed(1)

# Create a configuration
configuration <- ppjsdm::Configuration(x = 1:4, y = 2:5, marks = runif(4))

# Convert it to a data.frame
print(as.data.frame(configuration))

}
