% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{as.owin.Disk_window}
\alias{as.owin.Disk_window}
\title{Convert a disk window to an owin from the SpatStat package.}
\usage{
\method{as.owin}{Disk_window}(W, ..., fatal = TRUE)
}
\arguments{
\item{W}{Window.}

\item{...}{Unused.}

\item{fatal}{Unused.}
}
\description{
Convert a disk window to an owin from the SpatStat package.
}
\examples{
# Construct a window

window <- ppjsdm::Disk_window(centre = c(0, 2), radius = 2)

# Convert it to a spatstat object

print(as.owin(window))
}
