% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Im_window.R
\name{as.owin.Im_window}
\alias{as.owin.Im_window}
\title{Convert an im window to an owin from the \code{spatstat} package.}
\usage{
\method{as.owin}{Im_window}(W, ..., fatal = TRUE)
}
\arguments{
\item{W}{Window.}

\item{...}{Unused.}

\item{fatal}{Unused.}
}
\description{
Convert an im window to an owin from the \code{spatstat} package.
}
\examples{
if(require(spatstat.geom)) { # If no spatstat, im windows cannot be constructed
# Construct a window

window <- ppjsdm::Im_window(as.im(function(x, y) ifelse(x^2 + y^2 <= 1, 1, NA), W = owin()))

# Convert it to a spatstat object

print(as.owin(window))
}
}
