% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_generics.R
\name{as.owin}
\alias{as.owin}
\title{Convert Data To Class owin}
\usage{
as.owin(W, ..., fatal = TRUE)
}
\arguments{
\item{W}{Window.}

\item{...}{Ignored.}

\item{fatal}{What to do if the data cannot be converted to an observation window?}
}
\description{
Imported directly from \code{spatstat}, in order to provide methods for our window classes,
see for example \code{?as.owin.Rectangle_window}. The documentation here is sparse, and we
refer to the \code{spatstat} documentation for more details.
}
