% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window_union.R
\name{as.owin.Rectangle_window_union}
\alias{as.owin.Rectangle_window_union}
\title{Convert a union of rectangle windows to an owin from the \code{spatstat} package.}
\usage{
\method{as.owin}{Rectangle_window_union}(W, ..., fatal = TRUE)
}
\arguments{
\item{W}{Window.}

\item{...}{Unused.}

\item{fatal}{Unused.}
}
\description{
Convert a union of rectangle windows to an owin from the \code{spatstat} package.
}
\examples{
# Construct a window

window <- ppjsdm::Rectangle_window_union(x_ranges = list(c(0, 1), c(1, 2)), y_ranges = list(c(1, 3), c(2, 3)))

# Convert it to a spatstat object

print(as.owin(window))
plot(as.owin(window))
}
