% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{as.ppp.Configuration}
\alias{as.ppp.Configuration}
\title{Convert a configuration class to a ppp from the SpatStat package.}
\usage{
as.ppp.Configuration(X, W, ..., fatal = TRUE)
}
\arguments{
\item{X}{Configuration.}

\item{W}{Window on which the points are located.}

\item{...}{Not used.}

\item{fatal}{Not used.}
}
\description{
Convert a configuration class to a ppp from the SpatStat package.
}
