% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{as.Window}
\alias{as.Window}
\title{Convert another class to Window object.}
\usage{
as.Window(window)
}
\arguments{
\item{window}{Window.}
}
\description{
Convert another class to Window object.
}
\examples{
if(require(spatstat.geom)) { # If no spatstat, im windows cannot be constructed
# Construct an im

im <- as.im(function(x, y) ifelse(x^2 + y^2 <= 1, 1, NA), W = owin())

# Convert it to Window

window <- as.Window(im)

# Inspect it
print(window)
plot(window)
}
}
