% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_delta_phi_dispersion}
\alias{compute_delta_phi_dispersion}
\title{Compute the delta of the phi-dispersion of a marked configuration.}
\usage{
compute_delta_phi_dispersion(
  configuration,
  location,
  type,
  number_types,
  model = "identity",
  radius = matrix(1, 1)
)
}
\arguments{
\item{configuration}{Configuration.}

\item{location}{Point to be added.}

\item{type}{Type of point to be added.}

\item{number_types}{Number of different types.}

\item{model}{String representing the model to simulate from. You can check the currently authorised models with a call to `show_model()`.}

\item{radius}{Radius of interaction.}
}
\description{
Compute the delta of the phi-dispersion of a marked configuration.
}
