% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_papangelou.R
\name{compute_papangelou.default}
\alias{compute_papangelou.default}
\title{Compute the Papangelou conditional intensity of the model with given parameters.}
\usage{
\method{compute_papangelou}{default}(
  configuration,
  x,
  y,
  type = rep(1, length(x)),
  mark = rep(1, length(x)),
  alpha,
  gamma,
  beta0,
  covariates,
  beta,
  short_range,
  medium_range,
  long_range,
  saturation,
  types,
  model,
  medium_range_model,
  drop_type_from_configuration = FALSE,
  nthreads = 1,
  ...
)
}
\arguments{
\item{configuration}{Configuration.}

\item{x}{Coordinates along the x-axis of the points at which to evaluate the Papangelou conditional intensity.}

\item{y}{Coordinates along the y-axis of the points at which to evaluate the Papangelou conditional intensity.}

\item{type}{Type of the point (as an integer >= 1 or string representing the type).}

\item{mark}{Mark of the point to add.}

\item{alpha}{Matrix of short-range interaction coefficients.}

\item{gamma}{Matrix of medium-range interaction coefficients.}

\item{beta0}{Vector representing the intercept.}

\item{covariates}{List of covariates. These are converted to the `im` format by applying `as.im` to all elements in the list.}

\item{beta}{Fitted regression coefficients with respect to covariates.}

\item{short_range}{Symmetric matrix of short-range interaction radii. Can also be a list of matrices, each entry representing a different potential.}

\item{medium_range}{Symmetric matrix of medium-range interaction radii.}

\item{long_range}{Symmetric matrix of long-range interaction radii.}

\item{saturation}{Saturation parameter of the point process.}

\item{types}{Character vector, with entry i representing the name of type i.}

\item{model}{String representing the short-range model to use. The currently authorised models are obtained with a call to `show_short_range_models()`.}

\item{medium_range_model}{String representing the medium-range model to use. The currently authorised models are obtained with a call to `show_medium_range_models()`.}

\item{drop_type_from_configuration}{Should we remove the considered type(s) from the configuration?}

\item{nthreads}{Maximum number of threads for parallel computing.}

\item{...}{Ignored.}
}
\description{
Compute the Papangelou conditional intensity of the model with given parameters.
}
