% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_papangelou.R
\name{compute_papangelou.gibbsm}
\alias{compute_papangelou.gibbsm}
\title{Compute the Papangelou conditional intensity of the model from a fit object.}
\usage{
\method{compute_papangelou}{gibbsm}(
  fit,
  type = 1,
  configuration = fit$configuration_list[[1]],
  alpha = fit$coefficients$alpha,
  gamma = fit$coefficients$gamma,
  beta0 = fit$coefficients$beta0,
  covariates = fit$parameters$covariates,
  beta = fit$coefficients$beta,
  short_range = fit$coefficients$short_range,
  medium_range = fit$coefficients$medium_range,
  long_range = fit$coefficients$long_range,
  saturation = fit$parameters$saturation,
  types = fit$type_names,
  model = fit$parameters$model,
  medium_range_model = fit$parameters$medium_range_model,
  nthreads = fit$nthreads,
  ...
)
}
\arguments{
\item{fit}{Fit object obtained by running gibbsm.}

\item{type}{Type of the point at which to evaluate the Papangelou conditional intensity.}

\item{configuration}{Configuration of points at which to evaluate the Papangelou conditional intensity.}

\item{alpha}{Matrix of short-range interaction coefficients.}

\item{gamma}{Matrix of medium-range interaction coefficients.}

\item{beta0}{Vector representing the intercept.}

\item{covariates}{List of covariates. These are converted to the `im` format by applying `as.im` to all elements in the list.}

\item{beta}{Fitted regression coefficients with respect to covariates.}

\item{short_range}{Symmetric matrix of short-range interaction radii. Can also be a list of matrices, each entry representing a different potential.}

\item{medium_range}{Symmetric matrix of medium-range interaction radii.}

\item{long_range}{Symmetric matrix of long-range interaction radii.}

\item{saturation}{Saturation parameter of the point process.}

\item{types}{Character vector, with entry i representing the name of type i.}

\item{model}{String representing the short-range model to use. The currently authorised models are obtained with a call to `show_short_range_models()`.}

\item{medium_range_model}{String representing the medium-range model to use. The currently authorised models are obtained with a call to `show_medium_range_models()`.}

\item{nthreads}{Maximum number of threads for parallel computing.}

\item{...}{Forwarded to plot_papangelou}
}
\description{
Compute the Papangelou conditional intensity of the model from a fit object.
}
