% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_papangelou}
\alias{compute_papangelou}
\title{Compute Papangelou conditional intensity of the model.}
\usage{
compute_papangelou(
  configuration,
  x,
  y,
  type,
  model,
  medium_range_model,
  alpha,
  lambda,
  beta,
  gamma,
  covariates,
  short_range,
  medium_range,
  long_range,
  saturation
)
}
\arguments{
\item{configuration}{Configuration.}

\item{x}{Coordinates along the x-axis of the points at which to evaluate the Papangelou conditional intensity.}

\item{y}{Coordinates along the x-axis of the points at which to evaluate the Papangelou conditional intensity.}

\item{type}{Type of the point (as an integer >= 1).}

\item{model}{String representing the model to use. You can check the currently authorised models with a call to `show_models()`.}

\item{medium_range_model}{String representing the medium range model to use. You can check the currently authorised models with a call to `show_medium_range_models()`.}

\item{alpha}{Short range repulsion matrix.}

\item{lambda}{A vector representing the intensities of the point processes.}

\item{beta}{Coefficients related to covariates.}

\item{gamma}{Medium range repulsion matrix.}

\item{covariates}{Covariates.}

\item{short_range}{Short range interaction radii.}

\item{medium_range}{Medium range interaction radii.}

\item{long_range}{Long range interaction radii.}

\item{saturation}{Saturation parameter.}
}
\description{
Compute Papangelou conditional intensity of the model.
}
