% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_phi_dispersion}
\alias{compute_phi_dispersion}
\title{Compute the phi-dispersion of a marked configuration (radius = 0 corresponds to the proposal model).}
\usage{
compute_phi_dispersion(configuration, number_types, radius = 0)
}
\arguments{
\item{configuration}{The configuration.}

\item{number_types}{Number of different types.}

\item{radius}{Radius of interaction.}
}
\description{
Compute the phi-dispersion of a marked configuration (radius = 0 corresponds to the proposal model).
}
