% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_S.R
\name{compute_S}
\alias{compute_S}
\title{Calculate the matrix S used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This is an internal function used when you want to compute the vcov matrix in multiple steps (useful for large datasets).}
\usage{
compute_S(..., list, nthreads = NULL, debug = FALSE, time_limit = Inf)
}
\arguments{
\item{...}{A sequence of fit objects.}

\item{list}{List of fits.}

\item{nthreads}{(optional) number of threads to use.}

\item{debug}{Display debug information?}

\item{time_limit}{Time limit in hours that can be spent running this function.}
}
\description{
Calculate the matrix S used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This is an internal function used when you want to compute the vcov matrix in multiple steps (useful for large datasets).
}
