% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\docType{class}
\name{Configuration-class}
\alias{Configuration-class}
\title{A class representing a configuration.}
\arguments{
\item{x}{Values along the x-axis.}

\item{y}{Values along the y-axis.}

\item{types}{Types of the points.}
}
\description{
A class representing a configuration.
}
