% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{Configuration}
\alias{Configuration}
\title{Configuration constructor.}
\usage{
Configuration(x, y, types, marks)
}
\arguments{
\item{x}{Values along the x-axis.}

\item{y}{Values along the y-axis.}

\item{types}{Types of the points.}

\item{marks}{Marks of the points.}
}
\description{
Configuration constructor.
}
