% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{construct_if_missing}
\alias{construct_if_missing}
\title{Construct an object \code{x} that might already be constructed, but if not that should be constructed
using default supplied values, as either a matrix or a vector.}
\usage{
construct_if_missing(x, def, nrows, ncols = nrows, matrix = FALSE)
}
\arguments{
\item{x}{Object that contains either an object, or NULL if it should be constructed.}

\item{def}{Default value for the object.}

\item{nrows}{Number of rows of the object to construct.}

\item{ncols}{Number of columns of the object to construct.}

\item{matrix}{Should we construct a matrix or a vector?}
}
\description{
Construct an object \code{x} that might already be constructed, but if not that should be constructed
using default supplied values, as either a matrix or a vector.
}
\keyword{internal}
