% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{convert_names}
\alias{convert_names}
\title{Convert a vector of names to their corresponding entry in a named list.}
\usage{
convert_names(x, named_list, ...)
}
\arguments{
\item{x}{A vector of names.}

\item{named_list}{A named list, where names correspond to EITHER elements in x OR x converted according to one of the extra supplied objects
, and the value in \code{named_list} represent what we are converting to.}

\item{...}{We assume that the names in \code{named_list} might correspond to a conversion of x according to named lists in these objects.}
}
\description{
Convert a vector of names to their corresponding entry in a named list.
}
\keyword{internal}
