% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{create_list_of_fits}
\alias{create_list_of_fits}
\title{Construct a list of fits from a series of fit objects.}
\usage{
create_list_of_fits(..., list)
}
\arguments{
\item{...}{Any number of fit objects obtained by a call to \code{ppjsdm::gibbsm}.}

\item{list}{Some more fits provided as a list.}
}
\description{
Construct a list of fits from a series of fit objects.
}
\keyword{internal}
