% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coefficient_vector.R
\name{format_coefficient_vector}
\alias{format_coefficient_vector}
\title{Format the coefficient vector returned by gibbsm into vector and matrix form for all coefficients.}
\usage{
format_coefficient_vector(
  coefficient_vector,
  number_types,
  types_names,
  covariates_names,
  estimate_alpha,
  estimate_gamma
)
}
\arguments{
\item{coefficient_vector}{Coefficient vector to format.}

\item{number_types}{Number of types.}

\item{types_names}{Names of the types.}

\item{covariates_names}{Names of the covariates.}

\item{estimate_alpha}{Matrix where entry i, j is a boolean representing whether or not we are modelling short-range interactions between i and j.}

\item{estimate_gamma}{Matrix where entry i, j is a boolean representing whether or not we are modelling medium-range interactions between i and j.}
}
\description{
Format the coefficient vector returned by gibbsm into vector and matrix form for all coefficients.
}
