% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{get_marks}
\alias{get_marks}
\title{Access marks of a configuration}
\usage{
get_marks(configuration)
}
\arguments{
\item{configuration}{The configuration.}
}
\description{
Access marks of a configuration
}
