% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{get_number_points}
\alias{get_number_points}
\title{Number of points in a configuration}
\usage{
get_number_points(configuration)
}
\arguments{
\item{configuration}{The configuration.}
}
\description{
Number of points in a configuration
}
