% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{get_number_types_and_check_conformance}
\alias{get_number_types_and_check_conformance}
\title{Deduce from a list of objects the number of types the user has in mind.}
\usage{
get_number_types_and_check_conformance(default_number_types, types, ...)
}
\arguments{
\item{default_number_types}{Default number of types.}

\item{...}{Objects (usually matrices and vectors) that imply varying number of types.}
}
\description{
Deduce from a list of objects the number of types the user has in mind.
}
\keyword{internal}
