% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_from_fit.R
\name{get_parameters_from_fit}
\alias{get_parameters_from_fit}
\title{Get parameters from a fit.}
\usage{
get_parameters_from_fit(fit)
}
\arguments{
\item{fit}{Fit.}
}
\description{
Get parameters from a fit.
}
