% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qld_trees_configuration.R
\name{get_qld_trees_configuration}
\alias{get_qld_trees_configuration}
\title{Return a configuration representing the locations of 'ten' species of trees in a given CSIRO experimental plot index by 'index', in a given 'year'.}
\usage{
get_qld_trees_configuration(
  index = 15,
  year = 2013,
  prevalent = 10,
  jitter = 0.3,
  average_height = FALSE,
  average_dbh = FALSE
)
}
\arguments{
\item{index}{An index between 1 and 20 representing the experimental plot.}

\item{year}{A number between 1971 and 2013 representing the year.}

\item{prevalent}{Number of most prevalent species to use.}

\item{jitter}{Variance of the shifts applied to the locations of the trees.}

\item{average_height}{Return trait corresponding to average heights?}

\item{average_dbh}{Return trait corresponding to average stem diameters?}
}
\description{
Return a configuration representing the locations of 'ten' species of trees in a given CSIRO experimental plot index by 'index', in a given 'year'.
}
