% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsm.R
\name{gibbsm}
\alias{gibbsm}
\title{Fit a multivariate Gibbs model to a dataset.}
\usage{
gibbsm(
  configuration_list,
  window = Rectangle_window(),
  covariates = list(),
  traits = list(),
  model = "identity",
  radius = NULL,
  print = TRUE
)
}
\arguments{
\item{configuration_list}{A single configuration or a list of configurations assumed to be drawn from the multivariate Gibbs.}

\item{window}{Observation window.}

\item{covariates}{Environmental covariates driving the intensity.}

\item{traits}{Species' traits.}

\item{model}{String to represent the model we're calibrating. You can check the currently authorised models with a call to `show_model()`.}

\item{radius}{Interaction radius.}

\item{print}{Print the fitting summary?}
}
\description{
Fit a multivariate Gibbs model to a dataset.
}
