% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_duplicates}
\alias{has_duplicates}
\title{Check if a configuration contains duplicates.}
\usage{
has_duplicates(configuration)
}
\arguments{
\item{configuration}{Configuration.}
}
\description{
Check if a configuration contains duplicates.
}
