% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map.R
\name{heat_map}
\alias{heat_map}
\title{Heatmap of coefficient values of a fit object}
\usage{
heat_map(
  fit,
  coefficient = c("alpha", "beta"),
  show_values = TRUE,
  limits = NULL,
  include_y = NULL,
  include_x = NULL,
  names = NULL
)
}
\arguments{
\item{fit}{A fit object obtained by a call to 'ppjsdm::gibbsm'}

\item{coefficient}{A strong representing the coefficient to plot}

\item{show_values}{Should the cells show the coefficient value?}

\item{limits}{Optional vector that gives what coefficient values should be plotted}

\item{include_y}{Optional vector of types that should be included on the y-axis}

\item{include_x}{Optional vector of types that should be included on the x-axis}

\item{names}{Optional list of full names of types, if for example abbreviations were used when running the fit}
}
\description{
Heatmap of coefficient values of a fit object
}
\examples{
set.seed(1)

# Construct a configuration

configuration <- ppjsdm::rppp(lambda = c(A = 500, B = 500))

# Fit the model

fit <- ppjsdm::gibbsm(configuration, covariates = list(x = function(x, y) x))

# Plot the heatmap of coefficients

ppjsdm::heat_map(fit, coefficient = "x", show_values = FALSE)

}
