% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Im_window.R
\name{Im_window}
\alias{Im_window}
\title{Im window.}
\usage{
Im_window(im, window)
}
\arguments{
\item{im}{\code{spatstat} \code{im} object.}

\item{window}{Optional \code{spatstat} window object.}
}
\description{
Im window.
}
\examples{
if(require(spatstat.geom)) { # If no spatstat, im windows cannot be constructed
# Construct an im window

window <- ppjsdm::Im_window(as.im.function(function(x, y) ifelse(x^2 + y^2 <= 1, 1, NA), W = owin()))

print(window)

plot(window)
}
}
