% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{is_empty}
\alias{is_empty}
\title{Check whether a configuration is empty.}
\usage{
is_empty(configuration)
}
\arguments{
\item{configuration}{Configuration.}
}
\description{
Check whether a configuration is empty.
}
