% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_window.R
\name{is_in_window,ANY,Rectangle_window-method}
\alias{is_in_window,ANY,Rectangle_window-method}
\title{Check if points are in rectangle window}
\usage{
\S4method{is_in_window}{ANY,Rectangle_window}(points, window)
}
\arguments{
\item{points}{Points.}

\item{window}{Window}
}
\description{
Check if points are in rectangle window
}
