% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_window.R
\name{is_in_window}
\alias{is_in_window}
\title{Check if point is in window}
\usage{
is_in_window(x, y, window)
}
\arguments{
\item{x}{X-coordinates of points.}

\item{y}{Y-coordinates of points.}

\item{window}{Window.}
}
\description{
Check if point is in window
}
