% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_window.R
\name{is_in_window.Rectangle_window}
\alias{is_in_window.Rectangle_window}
\title{Check if points are in rectangle window}
\usage{
\method{is_in_window}{Rectangle_window}(x, y, window)
}
\arguments{
\item{x}{X-coordinates of points.}

\item{y}{Y-coordinates of points.}

\item{window}{Window}
}
\description{
Check if points are in rectangle window
}
