% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{length.Configuration}
\alias{length.Configuration}
\title{Number of points in a configuration}
\usage{
\method{length}{Configuration}(x)
}
\arguments{
\item{x}{Configuration.}
}
\description{
Number of points in a configuration
}
