% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{make_summary_df}
\alias{make_summary_df}
\title{Construct a \code{data.frame} containing estimated coefficients and their confidence intervals from some fit objects.}
\usage{
make_summary_df(
  fits,
  coefficient,
  summ,
  only_statistically_significant,
  which,
  full_names,
  compute_confidence_intervals,
  classes,
  involving,
  how
)
}
\arguments{
\item{fits}{A list of fit objected obtained by a call to \code{ppjsdm::gibbsm}.}

\item{coefficient}{A string representing the coefficient to plot.
Choice of \code{alpha1}, \code{alpha2}, ..., to show one of the short-range interaction coefficients;
\code{gamma} to show the medium-range interaction coefficient;
\code{beta1}, \code{beta2}, ... or actual name of the covariate for one of the regression coefficients.
\code{alpha} shows a facet plot of all short-range interaction coefficients, while \code{beta} does the same for the regression coefficients.}

\item{summ}{Optional list of summaries corresponding to the fits; if not provided they are obtained by calling \code{ppjsdm::summary.gibbsm}}

\item{only_statistically_significant}{Only show statistically significant coefficients?}

\item{which}{If plotting interaction coefficients, which ones do we want to plot?}

\item{full_names}{Optional list of full names of types, if for example abbreviations were used when running the fit.}

\item{compute_confidence_intervals}{Compute the confidence intervals (which is slower) or just show the point estimates?}

\item{classes}{If this parameter is supplied, then colours are used to distinguish classes instead of fits.
Only works when given a single fit.
Should be a named vector/list, with the names corresponding to types, and the value equal to the class name.}

\item{involving}{Optional vector/list of types. Only coefficients involving these types will be plotted.}

\item{how}{If the \code{involving} argument is supplied, should it involve \emph{only} those types, or at least \emph{one} of those types (which is relevant if inter-type interactions are involved).}
}
\description{
This function is used to prepare data for use in the \code{box_plot} and \code{diagram_plot} functions.
}
