% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{marks.Configuration}
\alias{marks.Configuration}
\title{Access marks of a configuration}
\usage{
\method{marks}{Configuration}(x, ...)
}
\arguments{
\item{x}{Configuration.}

\item{...}{Unused.}
}
\description{
Access marks of a configuration
}
\examples{
set.seed(1)

# Create a configuration
configuration <- ppjsdm::Configuration(x = 1:4, y = 2:5, marks = runif(4))

# Get its marks
print(marks(configuration))

}
