% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_generics.R
\name{marks}
\alias{marks}
\title{Marks of a Point Pattern}
\usage{
marks(x, ...)
}
\arguments{
\item{x}{Configuration of points.}

\item{...}{Ignored.}
}
\description{
Imported directly from \code{spatstat}, in order to provide a method for our Configuration class, see
\code{?marks.Configuration}. The documentation here is sparse, and we
refer to the \code{spatstat} documentation for more details.
}
