% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_papangelou.R
\name{model_matrix_papangelou}
\alias{model_matrix_papangelou}
\title{Compute the Papangelou conditional intensity of the model.}
\usage{
model_matrix_papangelou(...)
}
\arguments{
\item{...}{Parameters to be forwarded to the relevant method. Currently, either
the parameters of the Gibbs point process, or a fit object obtained from running `gibbsm`.}
}
\description{
Compute the Papangelou conditional intensity of the model.
}
