% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{plot.Configuration}
\alias{plot.Configuration}
\title{Plot a configuration class.}
\usage{
\method{plot}{Configuration}(x, window = NULL, ...)
}
\arguments{
\item{x}{Configuration to plot.}

\item{window}{Window the configuration belongs to.}

\item{...}{Other arguments not yet used.}
}
\description{
Plot a configuration class.
}
