% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{plot.gibbsm}
\alias{plot.gibbsm}
\title{Plot a \code{ppjsdm::gibbsm} object.}
\usage{
\method{plot}{gibbsm}(x, ...)
}
\arguments{
\item{x}{A fit object obtained by calling \code{ppjsdm::gibbsm}.}

\item{...}{Forwarded to \code{ppjsdm::box_plot}.}
}
\description{
This calls \code{ppjsdm::box_plot}, see that function's documentation for details.
}
