% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{plot.Window}
\alias{plot.Window}
\title{Plot a window.}
\usage{
\method{plot}{Window}(x, ...)
}
\arguments{
\item{x}{Window.}

\item{...}{Ignored.}
}
\description{
Plot a window.
}
\examples{
# Construct a window

window <- ppjsdm::Rectangle_window(x_range = c(0, 1), y_range = c(1, 3))

# Plot it

plot(window)
}
