% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_papangelou_intensity.R
\name{plot_papangelou.default}
\alias{plot_papangelou.default}
\title{Plot the Papangelou conditional intensity with given parameters.}
\usage{
\method{plot_papangelou}{default}(
  window,
  type = 1,
  mark = 1,
  configuration,
  alpha,
  gamma,
  beta0,
  covariates,
  beta,
  short_range,
  medium_range,
  long_range,
  saturation,
  types,
  model,
  medium_range_model,
  grid_steps = c(1000, 1000),
  steps = 0,
  nthreads = 1,
  use_log = FALSE,
  use_ggplot = TRUE,
  return_papangelou = FALSE,
  drop_type_from_configuration = FALSE,
  show = c("all", "none", "type"),
  limits,
  ...
)
}
\arguments{
\item{window}{Observation window. Preferably a \code{ppjsdm} Window, such as \code{ppjsdm::Rectangle_window}, but also accepts \code{spatstat} \code{im} or \code{owin} objects.}

\item{type}{Type of the point at which to evaluate the Papangelou conditional intensity. Either an integer representing the index of the type, or a string representing its label.}

\item{mark}{Mark of the point.}

\item{configuration}{Configuration of points at which to evaluate the Papangelou conditional intensity.}

\item{alpha}{Matrix of short-range interaction coefficients.}

\item{gamma}{Matrix of medium-range interaction coefficients.}

\item{beta0}{Vector representing the intercept.}

\item{covariates}{List of covariates. These are converted to the \code{im} format by applying \code{as.im} to all elements in the list.}

\item{beta}{Fitted regression coefficients with respect to covariates.}

\item{short_range}{Symmetric matrix of short-range interaction radii. Can also be a list of matrices, each entry representing a different potential.}

\item{medium_range}{Symmetric matrix of medium-range interaction radii.}

\item{long_range}{Symmetric matrix of long-range interaction radii.}

\item{saturation}{Saturation parameter of the point process.}

\item{types}{Character vector, with entry i representing the name of type i.}

\item{model}{String representing the short-range model to use. The currently authorised models are obtained with a call to \code{show_short_range_models()}.}

\item{medium_range_model}{String representing the medium-range model to use. The currently authorised models are obtained with a call to \code{show_medium_range_models()}.}

\item{grid_steps}{Vector of length 2, representing the number of horizontal and vertical grid steps. If a single value is given, it is assumed that the dimensions are the same along both axes.}

\item{steps}{Number of steps in the Metropolis-Hastings simulation algorithm. A value of 0 uses the CFTP algorithm.
This is only used if no configuration was supplied, in which case we simulate a configuration.}

\item{nthreads}{Maximum number of threads to use.}

\item{use_log}{Plot the logarithm of the Papangelou conditional intensity instead?}

\item{use_ggplot}{Use ggplot for fancier plot?}

\item{return_papangelou}{Should we return the Papangelou intensity as an \code{im} object instead of plotting it?}

\item{drop_type_from_configuration}{Should we remove the considered \code{type} from the configuration?}

\item{show}{In the plot, should we show all points, no points, or only the individuals of the configuration with the target type?}

\item{limits}{Limits for values of the Papangelou conditional intensity when plotting.}

\item{...}{Ignored.}
}
\description{
Plot the Papangelou conditional intensity with given parameters.
}
