% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_papangelou_intensity.R
\name{plot_papangelou.gibbsm}
\alias{plot_papangelou.gibbsm}
\title{Plot the Papangelou conditional intensity from a fit object.}
\usage{
\method{plot_papangelou}{gibbsm}(fit, ...)
}
\arguments{
\item{fit}{Fit object obtained by running gibbsm.}

\item{...}{Forwarded to plot_papangelou}
}
\description{
Plot the Papangelou conditional intensity from a fit object.
}
