% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials.R
\name{potentials}
\alias{potentials}
\title{Recover the interaction potential between two given species from a fitted `gibbsm` object.}
\usage{
potentials(fit, species1 = 1, species2 = species1)
}
\arguments{
\item{fit}{A `gibbsm` fit object.}

\item{species1}{First interacting species.}

\item{species2}{Second interacting species.}
}
\description{
Recover the interaction potential between two given species from a fitted `gibbsm` object.
}
