% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{print.Configuration}
\alias{print.Configuration}
\title{Print a configuration class.}
\usage{
\method{print}{Configuration}(x, ...)
}
\arguments{
\item{x}{Configuration to print.}

\item{...}{Other arguments not yet used.}
}
\description{
Print a configuration class.
}
