% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gibbsm.R
\name{print.summary_gibbsm}
\alias{print.summary_gibbsm}
\title{Print summary of a Fitted Gibbs point process.}
\usage{
\method{print}{summary_gibbsm}(x, ...)
}
\arguments{
\item{x}{A summary of a fitted model object.}

\item{...}{Ignored.}
}
\description{
Print summary of a Fitted Gibbs point process.
}
