% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbinomialpp.R
\name{rbinomialpp}
\alias{rbinomialpp}
\title{Sample a binomial point processes}
\usage{
rbinomialpp(
  window = Rectangle_window(),
  n = NULL,
  nsim = 1,
  types = NULL,
  drop = TRUE,
  mark_range = c(1, 1)
)
}
\arguments{
\item{window}{Observation window. Default is a rectangle window \eqn{[0, 1]^2}. Preferably a \code{ppjsdm} Window, such as \code{ppjsdm::Rectangle_window}, but also accepts \code{spatstat} \code{im} or \code{owin} objects.}

\item{n}{A vector representing the number of points of each types of the multipoint binomial point processes.
Default is a vector of same size as types, filled with ones.}

\item{nsim}{Number of samples to generate. Default is 1.}

\item{types}{Types of the points. Default is a vector (type1, type2, ...) of same size as n.}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration.
Default is TRUE.}

\item{mark_range}{Range of additional marks to give to the points.}
}
\description{
Sample a binomial point processes
}
