% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\docType{class}
\name{Rectangle_window-class}
\alias{Rectangle_window-class}
\title{A rectangle window}
\arguments{
\item{x_range}{Range along the x-axis.}

\item{y_range}{Range along the y-axis.}
}
\description{
A rectangle window
}
