% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{Rectangle_window}
\alias{Rectangle_window}
\title{Rectangle window constructor}
\usage{
Rectangle_window(x_range = c(0, 1), y_range = c(0, 1))
}
\arguments{
\item{x_range}{Range along the x-axis.}

\item{y_range}{Range along the y-axis.}
}
\description{
Rectangle window constructor
}
