% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window_union.R
\name{Rectangle_window_union}
\alias{Rectangle_window_union}
\title{Rectangle window union constructor}
\usage{
Rectangle_window_union(x_ranges = list(c(0, 1)), y_ranges = list(c(0, 1)))
}
\arguments{
\item{x_ranges}{Range along the x-axis.}

\item{y_ranges}{Range along the y-axis.}
}
\description{
Rectangle window union constructor
}
