% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{remove_from_configuration}
\alias{remove_from_configuration}
\title{Remove a point from configuration (if it is in the configuration).}
\usage{
remove_from_configuration(configuration, x, type)
}
\arguments{
\item{configuration}{Configuration.}

\item{x}{Point.}

\item{type}{Point type.}
}
\description{
Remove a point from configuration (if it is in the configuration).
}
