% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{remove_from_configuration_by_index}
\alias{remove_from_configuration_by_index}
\title{Remove a point from a configuration by index.}
\usage{
remove_from_configuration_by_index(configuration, index)
}
\arguments{
\item{configuration}{Configuration.}

\item{index}{Index.}
}
\description{
Remove a point from a configuration by index.
}
