% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{remove_random_point}
\alias{remove_random_point}
\title{Remove random point from the configuration.}
\usage{
remove_random_point(configuration)
}
\arguments{
\item{configuration}{Configuration.}
}
\description{
Remove random point from the configuration.
}
