% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rppp}
\alias{rppp}
\title{Sample a Poisson point processes}
\usage{
rppp(window, lambda = as.numeric(c(1)), nsim = 1L, types = NULL,
  drop = TRUE)
}
\arguments{
\item{window}{The window.}

\item{lambda}{A vector representing the intensities of the multipoint Poisson point processes.}

\item{nsim}{Number of samples to generate.}

\item{types}{Types of the points.}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration.}
}
\description{
Sample a Poisson point processes
}
