% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{show,Configuration-method}
\alias{show,Configuration-method}
\title{Print a configuration class.}
\usage{
\S4method{show}{Configuration}(object)
}
\arguments{
\item{object}{Configuration to print.}
}
\description{
Print a configuration class.
}
