% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{[.Configuration}
\alias{[.Configuration}
\title{Configuration subset operator.}
\usage{
\method{[}{Configuration}(x, i)
}
\arguments{
\item{x}{Configuration.}

\item{i}{Subscript index.}
}
\description{
Configuration subset operator.
}
