% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{[,Configuration,numeric,missing,ANY-method}
\alias{[,Configuration,numeric,missing,ANY-method}
\title{Configuration subset operator.}
\usage{
\S4method{[}{Configuration,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{Configuration.}

\item{i}{Subscript index.}

\item{j}{Ignored.}

\item{...}{Ignored.}

\item{drop}{Ignored.}
}
\description{
Configuration subset operator.
}
