% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Display a trace plot of the underlying Metropolis-Hastings algorithm used
to obtain a sample of a `ppjsdm` point process.}
\usage{
trace_plot(x, type = 1)
}
\arguments{
\item{x}{A simulated ppjsdm object, usually obtained from running `rgibbs` with the Metropolis-Hastings algorithm.}

\item{type}{Display trace plot corresponding to this type.}
}
\description{
Display a trace plot of the underlying Metropolis-Hastings algorithm used
to obtain a sample of a `ppjsdm` point process.
}
