% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.gibbsm.R
\name{vcov.gibbsm}
\alias{vcov.gibbsm}
\title{Calculate Variance-Covariance Matrix for a Fitted Gibbs point process.}
\usage{
\method{vcov}{gibbsm}(object, npoints = 2000, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{npoints}{Target number of points in the restricted window that the vcov matrix is computed on. Computation is slower for larger values, but the vcov matrix is then better approximated.}

\item{...}{Ignored.}
}
\description{
Calculate Variance-Covariance Matrix for a Fitted Gibbs point process.
}
