% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Im_window.R
\name{volume.Im_window}
\alias{volume.Im_window}
\title{Return the volume of an im window.}
\usage{
\method{volume}{Im_window}(x)
}
\arguments{
\item{x}{Window.}
}
\description{
Return the volume of an im window.
}
\examples{
if(require(spatstat.geom)) { # If no spatstat, im windows cannot be constructed
# Construct a window

window <- ppjsdm::Im_window(as.im(function(x, y) ifelse(x^2 + y^2 <= 1, 1, NA), W = owin()))

# Compute volume of an im window

print(volume(window))
}
}
