% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{volume.Rectangle_window}
\alias{volume.Rectangle_window}
\title{Return the volume of a rectangle window.}
\usage{
\method{volume}{Rectangle_window}(x)
}
\arguments{
\item{x}{The window.}
}
\description{
Return the volume of a rectangle window.
}
