% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\docType{methods}
\name{window_volume,Disk_window-method}
\alias{window_volume,Disk_window-method}
\title{Return the volume of the window.}
\usage{
\S4method{window_volume}{Disk_window}(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of the window.
}
