% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{window_volume,Rectangle_window-method}
\alias{window_volume,Rectangle_window-method}
\title{Return the volume of the window.}
\usage{
\S4method{window_volume}{Rectangle_window}(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of the window.
}
