% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{window_volume.Disk_window}
\alias{window_volume.Disk_window}
\title{Return the volume of the window.}
\usage{
\method{window_volume}{Disk_window}(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of the window.
}
