% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{window_volume}
\alias{window_volume}
\title{Return the volume of the window.}
\usage{
window_volume(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of the window.
}
