% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{window_volume.Rectangle_Window}
\alias{window_volume.Rectangle_Window}
\title{Return the volume of the window.}
\usage{
\method{window_volume}{Rectangle_Window}(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of the window.
}
