% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{x_coordinates}
\alias{x_coordinates}
\title{Access x-coordinates of a configuration}
\usage{
x_coordinates(configuration)
}
\arguments{
\item{configuration}{The configuration.}
}
\description{
Access x-coordinates of a configuration
}
