% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{x_range}
\alias{x_range}
\title{Access range along the x-axis of a rectangle window}
\usage{
x_range(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Access range along the x-axis of a rectangle window
}
