% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{y_coordinates}
\alias{y_coordinates}
\title{Access y-coordinates of a configuration}
\usage{
y_coordinates(configuration)
}
\arguments{
\item{configuration}{Configuration.}
}
\description{
Access y-coordinates of a configuration
}
