% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_range.R
\name{y_range}
\alias{y_range}
\title{Range along y-axis of enclosing rectangle of a window.}
\usage{
y_range(window)
}
\arguments{
\item{window}{Window.}
}
\description{
Range along y-axis of enclosing rectangle of a window.
}
\examples{
# Construct a disk window

window <- ppjsdm::Disk_window(centre = c(2, 1), radius = 2)

# Compute the range along the y-axis

print(ppjsdm::y_range(window))

}
