% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{area.Disk_window}
\alias{area.Disk_window}
\title{Return the area of a disk window.}
\usage{
\method{area}{Disk_window}(w)
}
\arguments{
\item{w}{Window.}
}
\description{
Return the area of a disk window.
}
\examples{
# Construct a window

window <- ppjsdm::Disk_window(centre = c(0, 2), radius = 2)

# Compute area of a rectangle window

print(area(window))
}
