% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{area.Disk_window}
\alias{area.Disk_window}
\title{Return the area of a disk window.}
\usage{
area.Disk_window(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the area of a disk window.
}
