% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{area.Rectangle_window}
\alias{area.Rectangle_window}
\title{Return the area of a rectangle window.}
\usage{
\method{area}{Rectangle_window}(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the area of a rectangle window.
}
