% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window_union.R
\name{area.Rectangle_window_union}
\alias{area.Rectangle_window_union}
\title{Return the area of a union of rectangle windows.}
\usage{
\method{area}{Rectangle_window_union}(w)
}
\arguments{
\item{w}{The window.}
}
\description{
Return the area of a union of rectangle windows.
}
