% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{as.Configuration}
\alias{as.Configuration}
\title{Convert to our configuration class.}
\usage{
as.Configuration(configuration)
}
\arguments{
\item{configuration}{Configuration.}
}
\description{
Convert to our configuration class.
}
