% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{as.owin.Rectangle_window}
\alias{as.owin.Rectangle_window}
\title{Convert a rectangle window to an owin from the \code{spatstat} package.}
\usage{
\method{as.owin}{Rectangle_window}(W, ..., fatal = TRUE)
}
\arguments{
\item{W}{Window.}

\item{...}{Unused.}

\item{fatal}{Unused.}
}
\description{
Convert a rectangle window to an owin from the \code{spatstat} package.
}
\examples{
# Construct a window

window <- ppjsdm::Rectangle_window(x_range = c(0, 1), y_range = c(1, 3))

# Convert it to a spatstat object

print(as.owin(window))
}
