% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_A2_plus_A3.R
\name{compute_A2_plus_A3}
\alias{compute_A2_plus_A3}
\title{Calculate the matrix A_2 + A3 used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This is an internal function used when you want to compute the vcov matrix in multiple steps (useful for large datasets).}
\usage{
compute_A2_plus_A3(
  ...,
  list,
  nthreads,
  debug = FALSE,
  npoints = 1000,
  multiple_windows = TRUE
)
}
\arguments{
\item{...}{A sequence of fit objects.}

\item{list}{List of fits.}

\item{nthreads}{(optional) number of threads to use.}

\item{debug}{Display debug information?}

\item{npoints}{Target number of points in the restricted window that the vcov matrix is computed on. Computation is slower for larger values, but the vcov matrix is then better approximated.}

\item{multiple_windows}{Compute A2 and A3 on a lot of small windows and which are then averaged out, or only on a single restricted window?}
}
\description{
Calculate the matrix A_2 + A3 used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This is an internal function used when you want to compute the vcov matrix in multiple steps (useful for large datasets).
}
