% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_G2.R
\name{compute_G2}
\alias{compute_G2}
\title{Calculate the matrix G2 used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This function is used to compute the vcov matrix in multiple steps (useful for large datasets).}
\usage{
compute_G2(
  ...,
  list,
  nthreads = NULL,
  debug = FALSE,
  time_limit = Inf,
  unit = "hours"
)
}
\arguments{
\item{...}{A sequence of fit objects.}

\item{list}{List of fits.}

\item{nthreads}{(optional) number of threads to use.}

\item{debug}{Display debug information?}

\item{time_limit}{Time limit measured in `unit` that can be spent running this function.}

\item{unit}{Unit used to measure the time limit (hours, mins, secs, etc).}
}
\description{
Calculate the matrix G2 used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This function is used to compute the vcov matrix in multiple steps (useful for large datasets).
}
