% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_papangelou}
\alias{compute_papangelou}
\title{Compute Papangelou conditional intensity of the model.}
\usage{
compute_papangelou(
  configuration,
  x,
  y,
  type,
  model,
  alpha,
  lambda,
  coefs,
  covariates,
  radius,
  saturation
)
}
\arguments{
\item{configuration}{Configuration.}

\item{x}{Coordinates along the x-axis of the points at which to evaluate the Papangelou conditional intensity.}

\item{y}{Coordinates along the x-axis of the points at which to evaluate the Papangelou conditional intensity.}

\item{type}{Type of the point (as an integer >= 1).}

\item{model}{String representing the model to simulate from. You can check the currently authorised models with a call to `show_model()`.}

\item{alpha}{Repulsion matrix.}

\item{lambda}{A vector representing the intensities of the point processes.}

\item{coefs}{Coefficients related to covariates.}

\item{covariates}{Covariates.}

\item{radius}{Interaction radii.}

\item{saturation}{Saturation parameter.}
}
\description{
Compute Papangelou conditional intensity of the model.
}
