% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_S.R
\name{compute_S}
\alias{compute_S}
\title{Calculate the matrix S used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This is an internal function used when you want to compute the vcov matrix in multiple steps (useful for large datasets).}
\usage{
compute_S(...)
}
\arguments{
\item{...}{A list of fit objects.}
}
\description{
Calculate the matrix S used to compute the Variance-Covariance Matrix for a Fitted Gibbs point process.
The function can be supplied multiple fits, in which case the aggregate estimator is computed.
This is an internal function used when you want to compute the vcov matrix in multiple steps (useful for large datasets).
}
