% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{Disk_window}
\alias{Disk_window}
\title{Disk window constructor}
\usage{
Disk_window(centre = c(0, 0), radius = 1)
}
\arguments{
\item{centre}{Centre of the disk.}

\item{radius}{Radius of the disk.}
}
\description{
Disk window constructor
}
