% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_until_time_limit.R
\name{execute_until_time_limit}
\alias{execute_until_time_limit}
\title{Execute a function on a list of objects, using educated guesses to execute as many calls
as possible within a specified time limit.}
\usage{
execute_until_time_limit(objects, func, time_limit = Inf, unit = "hours")
}
\arguments{
\item{objects}{List of objects.}

\item{func}{Function to apply to each object.}

\item{time_limit}{Time limit in `unit` that can be spent running this function.}

\item{unit}{Unit used to measure the time limit (hours, minutes, seconds, etc).}
}
\description{
Execute a function on a list of objects, using educated guesses to execute as many calls
as possible within a specified time limit.
}
\keyword{internal}
