% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{get_number_points}
\alias{get_number_points}
\title{Number of points in a configuration}
\usage{
get_number_points(configuration, total = FALSE)
}
\arguments{
\item{configuration}{Configuration.}

\item{total}{If set to `FALSE`, returns a list containing the number of points of the different types;
else returns the total number of points.}
}
\description{
Number of points in a configuration
}
