% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsm.R
\name{gibbsm}
\alias{gibbsm}
\title{Fit a multivariate Gibbs model to a dataset.}
\usage{
gibbsm(configuration, window, model = "identity", radius = 0, print = TRUE)
}
\arguments{
\item{configuration}{A configuration assumed to be a draw from the multivariate Gibbs.}

\item{window}{Observation window.}

\item{model}{String to represent the model we're calibrating. You can check the currently authorised models with a call to `show_model()`.}

\item{radius}{Interaction radius.}

\item{print}{Print the fitting summary?}
}
\description{
Fit a multivariate Gibbs model to a dataset.
}
