% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsm.R
\name{gibbsm}
\alias{gibbsm}
\title{Fit a multivariate Gibbs model to a dataset.}
\usage{
gibbsm(
  configuration_list,
  window = Rectangle_window(),
  covariates = list(),
  traits = list(),
  model = "square_bump",
  medium_range_model = "square_exponential",
  short_range = NULL,
  medium_range = NULL,
  long_range = NULL,
  saturation = 2,
  print = TRUE,
  use_glmnet = TRUE,
  use_aic = FALSE
)
}
\arguments{
\item{configuration_list}{A single configuration or a list of configurations assumed to be drawn from the multivariate Gibbs.}

\item{window}{Observation window.}

\item{covariates}{Environmental covariates driving the intensity.}

\item{traits}{Species' traits.}

\item{model}{String to represent the model we're calibrating. You can check the currently authorised models with a call to `show_short_range_models()`.}

\item{medium_range_model}{String to represent the model we're calibrating. You can check the currently authorised models with a call to `show_medium_range_models()`.}

\item{short_range}{Short range interaction radius.}

\item{medium_range}{Medium range interaction radius.}

\item{long_range}{Long range interaction radius.}

\item{saturation}{Saturation parameter of the point process.}

\item{print}{Print the fitted coefficients?}

\item{use_glmnet}{Use `glmnet` instead of `glm`?}

\item{use_aic}{Use AIC instead of BIC for model fitting?}
}
\description{
Fit a multivariate Gibbs model to a dataset.
}
