% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsm.R
\name{gibbsm}
\alias{gibbsm}
\title{Fit a multivariate Gibbs model to a dataset.}
\usage{
gibbsm(configuration, window, model = "i", radius = 0)
}
\arguments{
\item{configuration}{A configuration assumed to be a draw from the multivariate Gibbs.}

\item{window}{Observation window.}

\item{model}{String to represent the model we're calibrating. At the moment, either "i", "s" or "g".}

\item{radius}{Interaction radius.}
}
\description{
Fit a multivariate Gibbs model to a dataset.
}
