% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{model_parameters}
\alias{model_parameters}
\title{Construct parameters for the model.}
\usage{
model_parameters(
  window,
  alpha,
  gamma,
  beta0,
  covariates,
  beta,
  short_range,
  medium_range,
  long_range,
  saturation,
  types,
  model,
  medium_range_model,
  default_number_types = 1,
  ...
)
}
\arguments{
\item{window}{Simulation window.}

\item{alpha}{Repulsion matrix. Default is a square matrix of same size as types, filled with zeroes.}

\item{gamma}{Medium range repulsion matrix. Default is a square matrix of same size as types, filled with zeroes.}

\item{beta0}{A vector representing the log-intensities of the point processes.
Default is a vector of same size as types, filled with ones.}

\item{covariates}{Covariates, with an empty list as a default.}

\item{beta}{Fitted coefficients related to covariates. Default is square matrix of zeroes of the same
number of rows/columns as the covariates.}

\item{short_range}{Symmetric matrix of short range interaction radii. Filled with 0.1 by default.}

\item{medium_range}{Symmetric matrix of medium range interaction radii. Filled with 0 by default.}

\item{long_range}{Symmetric matrix of long range interaction radii. Filled with 0 by default.}

\item{saturation}{Saturation parameter of the point process. Default is 2.}

\item{types}{Types of the points. Default is a vector (type1, type2, ...) of same size as n.}

\item{model}{String representing the model to use You can check the currently authorised models with a call to `show_short_range_models()`.}

\item{medium_range_model}{String representing the model to use You can check the currently authorised models with a call to `show_medium_range_models()`.}

\item{default_number_types}{Default number of types.}

\item{...}{Other parameters used to infer the number of types.}
}
\description{
Construct parameters for the model.
}
