% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_papangelou_intensity.R
\name{plot_papangelou.default}
\alias{plot_papangelou.default}
\title{Plot the Papangelou conditional intensity with given parameters.}
\usage{
\method{plot_papangelou}{default}(
  window,
  type = 1,
  mark = 1,
  configuration,
  model,
  medium_range_model,
  alpha,
  beta0,
  beta,
  gamma,
  covariates,
  short_range,
  medium_range,
  long_range,
  types,
  saturation,
  grid_steps = c(1000, 1000),
  steps = 0,
  nthreads = 4,
  use_log = FALSE,
  use_ggplot = TRUE,
  return_papangelou = FALSE,
  drop_type_from_configuration = FALSE,
  show_only_type = FALSE,
  limits,
  ...
)
}
\arguments{
\item{window}{Observation window.}

\item{type}{Type of the point at which to evaluate the Papangelou conditional intensity.}

\item{mark}{Mark of the point.}

\item{configuration}{Configuration of points at which to evaluate the Papangelou conditional intensity.}

\item{model}{Model for short range interaction.}

\item{medium_range_model}{Model for medium range interaction.}

\item{alpha}{Short range repulsion parameter.}

\item{beta0}{Log-intensity.}

\item{beta}{Covariates coefficients.}

\item{gamma}{Medium range repulsion parameter.}

\item{covariates}{Covariates.}

\item{short_range}{Matrix of short range distances.}

\item{medium_range}{Matrix of medium range distances.}

\item{long_range}{Matrix of long range distances.}

\item{types}{Types of the points. Default is a vector (type1, type2, ...) of same size as n.}

\item{saturation}{Saturation.}

\item{grid_steps}{Vector of length 2, representing the number of horizontal and vertical grid steps.}

\item{steps}{Nunber of steps in the Metropolis-Hastings simulation algorithm.}

\item{nthreads}{Maximum number of threads for parallel computing.}

\item{use_log}{Plot the logarithm of the Papangelou conditional intensity instead?}

\item{use_ggplot}{Use ggplot for fancier plot?}

\item{return_papangelou}{Should we return the Papangelou intensity itself instead of plotting it?}

\item{drop_type_from_configuration}{Should we remove the considered type from the configuration?}

\item{show_only_type}{In the plot, should we show only the individuals of the configuration with the target type?}

\item{limits}{Limits for values of the Papamgelou conditional intensity in plotting functions.}

\item{...}{Ignored.}
}
\description{
Plot the Papangelou conditional intensity with given parameters.
}
