% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_papangelou_intensity.R
\name{plot_papangelou.gibbsm}
\alias{plot_papangelou.gibbsm}
\title{Plot the Papangelou conditional intensity from a fit object.}
\usage{
\method{plot_papangelou}{gibbsm}(
  fit,
  window = fit$window,
  configuration = fit$configuration_list[[1]],
  type = 1,
  alpha = fit$coefficients$alpha,
  gamma = fit$coefficients$gamma,
  beta0 = fit$coefficients$beta0,
  covariates = fit$parameters$covariates,
  beta = fit$coefficients$beta,
  short_range = fit$coefficients$short_range,
  medium_range = fit$coefficients$medium_range,
  long_range = fit$coefficients$long_range,
  saturation = fit$parameters$saturation,
  types = fit$type_names,
  model = fit$parameters$model,
  medium_range_model = fit$parameters$medium_range_model,
  nthreads = fit$nthreads,
  ...
)
}
\arguments{
\item{fit}{Fit object obtained by running gibbsm.}

\item{window}{Observation window. Preferably a \code{ppjsdm} Window, such as \code{ppjsdm::Rectangle_window}, but also accepts \code{spatstat} \code{im} or \code{owin} objects.}

\item{configuration}{Configuration of points at which to evaluate the Papangelou conditional intensity.}

\item{type}{Type of the point at which to evaluate the Papangelou conditional intensity. Either an integer representing the index of the type, or a string representing its label.}

\item{alpha}{Matrix of short-range interaction coefficients.}

\item{gamma}{Matrix of medium-range interaction coefficients.}

\item{beta0}{Vector representing the intercept.}

\item{covariates}{List of covariates. These are converted to the \code{im} format by applying \code{as.im} to all elements in the list.}

\item{beta}{Fitted regression coefficients with respect to covariates.}

\item{short_range}{Symmetric matrix of short-range interaction radii. Can also be a list of matrices, each entry representing a different potential.}

\item{medium_range}{Symmetric matrix of medium-range interaction radii.}

\item{long_range}{Symmetric matrix of long-range interaction radii.}

\item{saturation}{Saturation parameter of the point process.}

\item{types}{Character vector, with entry i representing the name of type i.}

\item{model}{String representing the short-range model to use. The currently authorised models are obtained with a call to \code{show_short_range_models()}.}

\item{medium_range_model}{String representing the medium-range model to use. The currently authorised models are obtained with a call to \code{show_medium_range_models()}.}

\item{nthreads}{Maximum number of threads to use.}

\item{...}{Forwarded to plot_papangelou}
}
\description{
Plot the Papangelou conditional intensity from a fit object.
}
