% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_papangelou_intensity.R
\name{plot_papangelou}
\alias{plot_papangelou}
\title{Plot the Papangelou conditional intensity.}
\usage{
plot_papangelou(
  window,
  type,
  configuration,
  model,
  medium_range_model,
  alpha,
  lambda,
  beta,
  gamma,
  covariates,
  short_range,
  medium_range,
  long_range,
  saturation,
  max_points,
  grid_steps = 1000,
  mark = 1,
  steps = 0
)
}
\arguments{
\item{window}{Observation window.}

\item{type}{Type of the point at which to evaluate the Papangelou conditional intensity.}

\item{configuration}{Configuration of points at which to evaluate the Papangelou conditional intensity.}

\item{model}{Model for short range interaction.}

\item{medium_range_model}{Model for medium range interaction.}

\item{alpha}{Short range repulsion parameter.}

\item{lambda}{Intensity.}

\item{beta}{Covariates coefficients.}

\item{gamma}{Medium range repulsion parameter.}

\item{covariates}{Covariates.}

\item{short_range}{Matrix of short range distances.}

\item{medium_range}{Matrix of medium range distances.}

\item{long_range}{Matrix of long range distances.}

\item{saturation}{Saturation.}

\item{max_points}{Maximum number of points.}

\item{grid_steps}{Number of horizontal and vertical grid steps.}

\item{mark}{Mark of the point.}

\item{steps}{Nunber of steps in the Metropolis-Hastings simulation algorithm.}
}
\description{
Plot the Papangelou conditional intensity.
}
