% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_papangelou_intensity.R
\name{plot_papangelou}
\alias{plot_papangelou}
\title{Plot the Papangelou conditional intensity.}
\usage{
plot_papangelou(...)
}
\arguments{
\item{...}{Parameters to be forwarded to the relevant method. Currently, either
the parameters of the Gibbs point process, or a fit object obtained from running `gibbsm`.}
}
\description{
IMPORTANT: Check ?plot_papangelou.default for the documentation.
}
