% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbinomialpp}
\alias{rbinomialpp}
\title{Sample a binomial point processes}
\usage{
rbinomialpp(window, n = as.integer(c(1)), nsim = 1L, types = NULL, drop = TRUE)
}
\arguments{
\item{window}{The window.}

\item{n}{A vector representing the number of points of each types of the multipoint binomial point processes.}

\item{nsim}{Number of samples to generate.}

\item{types}{Types of the points.}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration.}
}
\description{
Sample a binomial point processes
}
