% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbinomialpp}
\alias{rbinomialpp}
\title{Sample a binomial point processes}
\usage{
rbinomialpp(window = NULL, n = NULL, nsim = 1L, types = NULL,
  drop = TRUE)
}
\arguments{
\item{window}{Simulation window. Default is a Rectangle window [0, 1]^2.}

\item{n}{A vector representing the number of points of each types of the multipoint binomial point processes.
Default is a vector of same size as types, filled with ones.}

\item{nsim}{Number of samples to generate. Default is 1.}

\item{types}{Types of the points. Default is a vector (type1, type2, ...) of same size as n.}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration.
Default is TRUE.}
}
\description{
Sample a binomial point processes
}
