% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{remove_from_configuration}
\alias{remove_from_configuration}
\title{Remove a point from a configuration (if it is in the configuration).}
\usage{
remove_from_configuration(configuration, x, type = "default", mark = 1)
}
\arguments{
\item{configuration}{Configuration.}

\item{x}{Point.}

\item{type}{Point type.}

\item{mark}{Point mark.}
}
\description{
Remove a point from a configuration (if it is in the configuration).
}
