% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgibbs.R
\name{rgibbs.gibbsm}
\alias{rgibbs.gibbsm}
\title{Sample a multivariate Gibbs point processes from a fit object.}
\usage{
\method{rgibbs}{gibbsm}(fit, ...)
}
\arguments{
\item{fit}{Fit object obtained by running gibbsm.}

\item{...}{Forwarded to rgibbs.}
}
\description{
Sample a multivariate Gibbs point processes from a fit object.
}
