% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgibbs.R
\name{rgibbs}
\alias{rgibbs}
\title{Sample from a multivariate Gibbs point process.}
\usage{
rgibbs(
  alpha,
  lambda,
  nsim = 30000,
  prob = 0.5,
  verbose = TRUE,
  return_full_chain = FALSE
)
}
\arguments{
\item{alpha}{alpha}

\item{lambda}{lambda}
}
\description{
Sample from a multivariate Gibbs point process.
}
