% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgibbs.R
\name{rgibbs}
\alias{rgibbs}
\title{Sample a multivariate saturated pairwise interaction Gibbs point processes.
IMPORTANT: Check ?rgibbs.default for the documentation.}
\usage{
rgibbs(...)
}
\arguments{
\item{...}{Parameters to be forwarded to the relevant method. Currently, either
the parameters of the Gibbs point process, or a fit object obtained from running `gibbsm`.}
}
\description{
Sample a multivariate saturated pairwise interaction Gibbs point processes.
IMPORTANT: Check ?rgibbs.default for the documentation.
}
