% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgibbs.R
\name{rgibbs}
\alias{rgibbs}
\title{Sample a multivariate Gibbs point processes}
\usage{
rgibbs(
  window = NULL,
  alpha = NULL,
  lambda = NULL,
  covariates = NULL,
  coefs = NULL,
  radius = NULL,
  steps = 0,
  nsim = 1,
  types = NULL,
  model = "identity",
  drop = TRUE
)
}
\arguments{
\item{window}{Simulation window.}

\item{alpha}{Repulsion matrix. Default is a square matrix of same size as types, filled with zeroes.}

\item{lambda}{A vector representing the intensities of the point processes.
Default is a vector of same size as types, filled with ones.}

\item{covariates}{Covariates, with an empty list as a default.}

\item{coefs}{Fitted coefficients related to covariates. Default is square matrix of zeroes of the same
number of rows/columns as the covariates.}

\item{radius}{Symmetric matrix of interaction radii. Filled by zeroes by default;}

\item{steps}{Number of steps in the Metropolis algorithm. If `steps = 0`, uses the coupling from the past algorithm instead.
Default is 0.}

\item{nsim}{Number of samples to generate. Default is 1.}

\item{types}{Types of the points. Default is a vector (type1, type2, ...) of same size as n.}

\item{model}{String representing the model to simulate from. You can check the currently authorised models with a call to `show_model()`.}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration. Default is TRUE.}
}
\description{
Sample a multivariate Gibbs point processes
}
