% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultigibbs}
\alias{rmultigibbs}
\title{Sample a multivariate Gibbs point processes}
\usage{
rmultigibbs(
  window,
  alpha = 1L,
  lambda = 1L,
  radius = 0,
  steps = 30000L,
  nsim = 1L,
  types = NULL,
  model = "i",
  drop = TRUE
)
}
\arguments{
\item{window}{The window.}

\item{alpha}{Alpha.}

\item{lambda}{A vector representing the intensities of the point processes.}

\item{radius}{Interaction radius.}

\item{steps}{Number of steps in the Metropolis algorithm.}

\item{nsim}{Number of samples to generate.}

\item{types}{Types of the points.}

\item{model}{String representing the model to simulate from. At the moment, either "i", "s" or "g".}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration.}
}
\description{
Sample a multivariate Gibbs point processes
}
