% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppp.R
\name{rppp}
\alias{rppp}
\title{Sample a Poisson point processes}
\usage{
rppp(
  window = Rectangle_window(),
  lambda = NULL,
  nsim = 1,
  types = NULL,
  drop = TRUE,
  mark_range = c(1, 1)
)
}
\arguments{
\item{window}{Simulation window. Default is a Rectangle window [0, 1]^2.}

\item{lambda}{A vector representing the intensities of the multipoint Poisson point processes.
Default is a vector of same size as types, filled with 100s.}

\item{nsim}{Number of samples to generate. Default is 1.}

\item{types}{Types of the points. Default is a vector (type1, type2, ...) of same size as n.}

\item{drop}{If nsim = 1 and drop = TRUE, the result will be a Configuration, rather than a list containing a Configuration.
Default is TRUE.}

\item{mark_range}{Range of additional marks to give to the points.}
}
\description{
Sample a Poisson point processes
}
