% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{[.Configuration}
\alias{[.Configuration}
\title{Subset of the configuration consisting only of points of the i-th type.}
\usage{
\method{[}{Configuration}(x, i)
}
\arguments{
\item{x}{Configuration.}

\item{i}{Subscript index.}
}
\description{
Subset of the configuration consisting only of points of the i-th type.
}
