% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gibbsm.R
\name{summary.gibbsm}
\alias{summary.gibbsm}
\title{Summary of a Fitted Gibbs point process.}
\usage{
\method{summary}{gibbsm}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Ignored.}
}
\description{
Summary of a Fitted Gibbs point process.
}
