% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.gibbsm.R
\name{vcov.gibbsm}
\alias{vcov.gibbsm}
\title{Calculate Variance-Covariance Matrix for a Fitted Gibbs point process.}
\usage{
\method{vcov}{gibbsm}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Ignored.}
}
\description{
Calculate Variance-Covariance Matrix for a Fitted Gibbs point process.
}
