% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disk_window.R
\name{volume.Disk_window}
\alias{volume.Disk_window}
\title{Return the volume of a disk window.}
\usage{
\method{volume}{Disk_window}(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of a disk window.
}
