% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{volume.Rectangle_window}
\alias{volume.Rectangle_window}
\title{Return the volume of a rectangle window.}
\usage{
\method{volume}{Rectangle_window}(x)
}
\arguments{
\item{x}{Window.}
}
\description{
Return the volume of a rectangle window.
}
\examples{
# Construct a window

window <- ppjsdm::Rectangle_window(x_range = c(0, 1), y_range = c(1, 3))

# Compute volume of a rectangle window

print(volume(window))
}
