% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{volume.Rectangle_window}
\alias{volume.Rectangle_window}
\title{Return the volume of a rectangle window.}
\usage{
volume.Rectangle_window(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Return the volume of a rectangle window.
}
