% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window_union.R
\name{volume.Rectangle_window_union}
\alias{volume.Rectangle_window_union}
\title{Return the volume of a rectangle window.}
\usage{
\method{volume}{Rectangle_window_union}(x)
}
\arguments{
\item{x}{The window.}
}
\description{
Return the volume of a rectangle window.
}
