% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_range.R
\name{x_range}
\alias{x_range}
\title{Range along x-axis of enclosing rectangle of window.}
\usage{
x_range(window)
}
\arguments{
\item{window}{Window.}
}
\description{
Range along x-axis of enclosing rectangle of window.
}
