% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle_window.R
\name{y_range}
\alias{y_range}
\title{Access range along the y-axis of a rectangle window}
\usage{
y_range(window)
}
\arguments{
\item{window}{The window.}
}
\description{
Access range along the y-axis of a rectangle window
}
