% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_range.R
\name{y_range}
\alias{y_range}
\title{Range along y-axis of enclosing rectangle of window.}
\usage{
y_range(window)
}
\arguments{
\item{window}{Window.}
}
\description{
Range along y-axis of enclosing rectangle of window.
}
